perf: massive performance boons
The bottleneck in terms of performance with this
page is the fragment shader. The entire metaballs
render pipeline happens in the fragment shader, so
each pixel rendered incurs a big cost.

To circumvent this, I render the metaballs on a render
target that is 1/12th the size of the viewport. The
resultant render is send as a texture to the main
cavnas and used to determine if that pixel needs to be
rendered or not. This means that pixels with no metaball
intersections are largely discarded.